Update README, artist model, urls, and track views
Update README to reflect the state of the project.
Update the artist model to print only the name of the artist rather than a full sentence.
Update the urls with missing carrot to properly CRUD on tracks.
Update track views to pass in the correctly formatted object for a 'PATCH'